refactor: split data loading and label handling out of promql engine.rs
Pure code move, no behavior change:
- label_usage.rs: skip-labels AST analysis (labels_dropped_at_root)
- series_labels.rs: label attach phase, with the cache hit path
  (attach_cached_labels) and the miss scan (scan_missing_labels)
  as separate functions
- selector_loader.rs: selector_load_data_from_datafusion and the
  sample/exemplar loaders
