+++ src/service/promql/engine.rs
@@ -15,29 +15,10 @@
 
 use std::{str::FromStr, sync::Arc, time::Duration};
 
-use arrow::array::Array;
 use async_recursion::async_recursion;
-use config::{
-    TIMESTAMP_COL_NAME,
-    meta::promql::{EXEMPLARS_LABEL, HASH_LABEL, NAME_LABEL, VALUE_LABEL, value::*},
-    utils::{
-        hash::{Sum64, gxhash},
-        json,
-    },
-};
-use datafusion::{
-    arrow::{
-        array::{Float64Array, Int64Array, StringArray, UInt64Array},
-        datatypes::{DataType, Schema},
-    },
-    error::{DataFusionError, Result},
-    logical_expr::utils::disjunction,
-    physical_plan::{
-        Partitioning, execute_stream_partitioned, expressions::Column, repartition::RepartitionExec,
-    },
-    prelude::{DataFrame, Expr, SessionContext, col, lit},
-};
-use futures::{TryStreamExt, future::try_join_all};
+use config::meta::promql::{NAME_LABEL, value::*};
+use datafusion::error::{DataFusionError, Result};
+use futures::future::try_join_all;
 use hashbrown::{HashMap, HashSet};
 use infra::errors::{Error, ErrorCodes};
 use promql_parser::{
@@ -48,29 +29,17 @@ use promql_parser::{
         VectorMatchCardinality, VectorSelector, token,
     },
 };
-use rayon::iter::{
-    IntoParallelIterator, IntoParallelRefIterator, IntoParallelRefMutIterator, ParallelIterator,
-};
+use rayon::iter::{IntoParallelIterator, IntoParallelRefMutIterator, ParallelIterator};
 
 use super::{
-    PromqlContext, label_cache,
-    utils::{apply_label_selector, apply_matchers},
+    PromqlContext, label_usage::labels_dropped_at_root,
+    selector_loader::selector_load_data_from_datafusion,
 };
 use crate::service::promql::{
     aggregations, binaries, functions, micros, rewrite::remove_filter_all,
 };
 #[cfg(feature = "enterprise")]
 use crate::service::search::SEARCH_SERVER;
-
-type TokioResult = tokio::task::JoinHandle<Result<HashMap<u64, Vec<Sample>>>>;
-type TokioExemplarsResult = tokio::task::JoinHandle<Result<HashMap<u64, Vec<Arc<Exemplar>>>>>;
-
-// Constants for optimization thresholds
-const OPTIMIZATION_STEP_LOOKBACK_MULTIPLIER: i64 = 5;
-const OPTIMIZATION_MAX_STEPS: i64 = 30;
-// Upper bound on the number of series hashes put into an in-list filter when
-// narrowing the label scan to cache-missed series.
-const MAX_HASH_INLIST_FILTER: usize = 8192;
 
 pub struct Engine {
     trace_id: String,
@@ -1288,650 +1257,6 @@ impl Engine {
     }
 }
 
-#[allow(clippy::too_many_arguments)]
-async fn selector_load_data_from_datafusion(
-    query_ctx: Arc<QueryContext>,
-    ctx: SessionContext,
-    schema: Arc<Schema>,
-    selector: VectorSelector,
-    label_selector: HashSet<String>,
-    start: i64,
-    end: i64,
-    step: i64,
-    lookback: i64,
-    skip_labels: bool,
-) -> Result<HashMap<u64, RangeValue>> {
-    let start_time = std::time::Instant::now();
-    let table_name = selector.name.as_ref().unwrap();
-
-    let mut df_group = match ctx.table(table_name).await {
-        Ok(v) => {
-            // Optimization: When step > lookback, we don't need to load all data in
-            // [start-lookback, end] Instead, we only need to load data windows around
-            // each evaluation point
-            let use_optimization = start != end
-                && step > 0
-                && step >= lookback * OPTIMIZATION_STEP_LOOKBACK_MULTIPLIER
-                && (((end - start) / step) + 1) < OPTIMIZATION_MAX_STEPS;
-            if use_optimization {
-                let num_steps = ((end - start) / step) + 1;
-                let eval_timestamps: Vec<i64> =
-                    (0..num_steps).map(|i| start + (step * i)).collect();
-
-                let mut conditions: Vec<Expr> = Vec::new();
-                for &eval_ts in &eval_timestamps {
-                    let window_start = eval_ts - lookback;
-                    let window_end = eval_ts;
-
-                    conditions.push(
-                        col(TIMESTAMP_COL_NAME)
-                            .gt_eq(lit(window_start))
-                            .and(col(TIMESTAMP_COL_NAME).lt_eq(lit(window_end))),
-                    );
-                }
-
-                let filters = disjunction(conditions).unwrap();
-                v.filter(filters)?
-            } else {
-                // Need to include lookback window before start for the first evaluation point
-                let query_start = start - lookback;
-                v.filter(
-                    col(TIMESTAMP_COL_NAME)
-                        .gt_eq(lit(query_start))
-                        .and(col(TIMESTAMP_COL_NAME).lt_eq(lit(end))),
-                )?
-            }
-        }
-        Err(_) => {
-            return Ok(HashMap::default());
-        }
-    };
-
-    df_group = apply_matchers(df_group, &schema, &selector.matchers)?;
-
-    match apply_label_selector(df_group, &schema, &label_selector) {
-        Some(dataframe) => df_group = dataframe,
-        None => return Ok(HashMap::default()),
-    }
-
-    // check if exemplars field is exists
-    if query_ctx.query_exemplars {
-        let schema = df_group.schema().as_arrow();
-        if schema.field_with_name(EXEMPLARS_LABEL).is_err() {
-            return Ok(HashMap::default());
-        }
-    }
-
-    // get label columns
-    let mut label_col_names = df_group
-        .schema()
-        .fields()
-        .iter()
-        .filter_map(|field| {
-            let name = field.name();
-            if name == TIMESTAMP_COL_NAME || name == VALUE_LABEL || name == EXEMPLARS_LABEL {
-                None
-            } else {
-                Some(name.to_string())
-            }
-        })
-        .collect::<Vec<_>>();
-    // sort labels to have a consistent order
-    label_col_names.sort();
-
-    // get hash & timestamp
-    let start1 = std::time::Instant::now();
-    let hash_field_type = schema.field_with_name(HASH_LABEL)?.data_type();
-    let mut metrics = if query_ctx.query_exemplars {
-        load_exemplars_from_datafusion(&query_ctx.trace_id, hash_field_type, df_group.clone())
-            .await?
-    } else {
-        load_samples_from_datafusion(&query_ctx.trace_id, hash_field_type, df_group.clone()).await?
-    };
-
-    log::info!(
-        "[trace_id: {}] load hashing and sample took: {:?}, metrics count: {}",
-        query_ctx.trace_id,
-        start1.elapsed(),
-        metrics.len(),
-    );
-
-    // The query provably discards all labels (e.g. `sum(rate(m[5m]))`), so
-    // the label scan / cache phase can be skipped entirely.
-    if skip_labels {
-        log::info!(
-            "[trace_id: {}] skip loading labels: query drops all labels",
-            query_ctx.trace_id,
-        );
-        return Ok(metrics);
-    }
-
-    // Labels are immutable per series hash: serve them from the cache and
-    // scan the label columns only for series that are not cached yet.
-    let start2 = std::time::Instant::now();
-    let label_cache = &*label_cache::LABEL_CACHE;
-    // bypass the cache when the working set won't fit (stored labels
-    // exclude the hash column)
-    let cache_enabled = label_cache.admit(label_col_names.len().saturating_sub(1), metrics.len());
-    if !cache_enabled {
-        log::info!(
-            "[trace_id: {}] label cache bypassed: {} series x {} label cols exceeds the cache budget",
-            query_ctx.trace_id,
-            metrics.len(),
-            label_col_names.len(),
-        );
-    }
-    let ctx_fp = label_cache::context_fingerprint(&query_ctx.org_id, table_name, &label_col_names);
-    // attach cached labels in parallel; millions of series make the
-    // lookup+clone loop CPU-bound
-    let missing_hashes: Vec<u64> = if cache_enabled {
-        metrics
-            .iter_mut()
-            .collect::<Vec<_>>()
-            .into_par_iter()
-            .filter_map(|(hash, range_val)| match label_cache.get(ctx_fp, *hash) {
-                Some(labels) => {
-                    range_val.labels =
-                        maybe_prepend_hash_label(labels, *hash, query_ctx.query_data);
-                    None
-                }
-                None => Some(*hash),
-            })
-            .collect()
-    } else {
-        metrics.keys().copied().collect()
-    };
-    let cache_hits = metrics.len() - missing_hashes.len();
-    config::metrics::QUERY_METRICS_LABEL_CACHE_HIT_COUNT
-        .with_label_values(&[&query_ctx.org_id])
-        .inc_by(cache_hits as u64);
-    config::metrics::QUERY_METRICS_LABEL_CACHE_MISS_COUNT
-        .with_label_values(&[&query_ctx.org_id])
-        .inc_by(missing_hashes.len() as u64);
-
-    if !missing_hashes.is_empty() {
-        // membership test for the extraction loop; skip the set when every
-        // series missed (cache bypassed or cold)
-        let all_missing = missing_hashes.len() == metrics.len();
-        let missing_set: HashSet<u64> = if all_missing {
-            HashSet::new()
-        } else {
-            missing_hashes.iter().copied().collect()
-        };
-        // hashes whose labels were extracted in this scan (dedup)
-        let mut labeled_set: HashSet<u64> = HashSet::new();
-        // Each missing series has a row at its own max timestamp, so scanning
-        // those timestamps yields at least one label row per missing series.
-        let series_timestamps = missing_hashes
-            .par_iter()
-            .filter_map(|hash| {
-                let range_val = metrics.get(hash)?;
-                let sample_max = range_val.samples.iter().map(|s| s.timestamp).max();
-                let exemplar_max = range_val
-                    .exemplars
-                    .as_ref()
-                    .and_then(|v| v.iter().map(|e| e.timestamp).max());
-                sample_max.max(exemplar_max)
-            })
-            .collect::<Vec<_>>();
-
-        let mut df_series = df_group;
-        // narrow the scan to the missing series when the hash column allows
-        // building an in-list filter of reasonable size
-        if hash_field_type == &DataType::UInt64 && missing_hashes.len() <= MAX_HASH_INLIST_FILTER {
-            df_series = df_series.filter(col(HASH_LABEL).in_list(
-                missing_hashes.iter().map(|&v| lit(v)).collect::<Vec<_>>(),
-                false,
-            ))?;
-        }
-        let label_cols = label_col_names
-            .iter()
-            .map(|name| col(name.as_str()))
-            .collect::<Vec<_>>();
-        let series = if config::get_config().limit.metrics_inlist_filter_enabled
-            || series_timestamps.is_empty()
-        {
-            // many series share the same max timestamp; dedup to keep the
-            // in-list filter small
-            let mut in_list_timestamps = series_timestamps;
-            in_list_timestamps.sort_unstable();
-            in_list_timestamps.dedup();
-            df_series
-                .filter(
-                    col(TIMESTAMP_COL_NAME).in_list(
-                        in_list_timestamps
-                            .iter()
-                            .map(|&v| lit(v))
-                            .collect::<Vec<_>>(),
-                        false,
-                    ),
-                )?
-                .select(label_cols)?
-                .collect()
-                .await?
-        } else {
-            let min = series_timestamps.iter().min().unwrap();
-            let max = series_timestamps.iter().max().unwrap();
-            df_series
-                .filter(col(TIMESTAMP_COL_NAME).between(lit(*min), lit(*max)))?
-                .select(label_cols)?
-                .collect()
-                .await?
-        };
-
-        for batch in series {
-            let columns = batch.columns();
-            let schema = batch.schema();
-            let fields = schema.fields();
-            let cols = fields
-                .iter()
-                .zip(columns)
-                .filter_map(|(field, col)| {
-                    if field.name() == HASH_LABEL {
-                        None
-                    } else {
-                        col.as_any()
-                            .downcast_ref::<StringArray>()
-                            .map(|col| (field.name(), col))
-                    }
-                })
-                .collect::<Vec<(_, _)>>();
-            let mut attach_row_labels = |hash: u64, i: usize| {
-                if (!all_missing && !missing_set.contains(&hash)) || labeled_set.contains(&hash) {
-                    return;
-                }
-                let Some(range_val) = metrics.get_mut(&hash) else {
-                    return;
-                };
-                let mut labels = Vec::with_capacity(cols.len());
-                for (name, value) in cols.iter() {
-                    if value.is_null(i) {
-                        continue;
-                    }
-                    labels.push(Arc::new(Label {
-                        name: name.to_string(),
-                        value: value.value(i).to_string(),
-                    }));
-                }
-                labeled_set.insert(hash);
-                if cache_enabled {
-                    label_cache.put(ctx_fp, hash, labels.clone());
-                }
-                range_val.labels = maybe_prepend_hash_label(labels, hash, query_ctx.query_data);
-            };
-            if hash_field_type == &DataType::UInt64 {
-                let hash_values = batch
-                    .column_by_name(HASH_LABEL)
-                    .unwrap()
-                    .as_any()
-                    .downcast_ref::<UInt64Array>()
-                    .unwrap();
-                for i in 0..batch.num_rows() {
-                    attach_row_labels(hash_values.value(i), i);
-                }
-            } else {
-                let hash_values = batch
-                    .column_by_name(HASH_LABEL)
-                    .unwrap()
-                    .as_any()
-                    .downcast_ref::<StringArray>()
-                    .unwrap();
-                for i in 0..batch.num_rows() {
-                    attach_row_labels(gxhash::new().sum64(hash_values.value(i)), i);
-                }
-            }
-        }
-    }
-
-    log::info!(
-        "[trace_id: {}] load all labels took: {:?}, label cache hits: {}, misses: {}",
-        query_ctx.trace_id,
-        start2.elapsed(),
-        cache_hits,
-        missing_hashes.len(),
-    );
-
-    log::info!(
-        "[trace_id: {}] load data from datafusion took: {:?}",
-        query_ctx.trace_id,
-        start_time.elapsed(),
-    );
-
-    Ok(metrics)
-}
-
-async fn load_samples_from_datafusion(
-    trace_id: &str,
-    hash_field_type: &DataType,
-    df: DataFrame,
-) -> Result<HashMap<u64, RangeValue>> {
-    let ctx = Arc::new(df.task_ctx());
-    let target_partitions = ctx.session_config().target_partitions();
-    let plan = df
-        .select_columns(&[TIMESTAMP_COL_NAME, HASH_LABEL, VALUE_LABEL])?
-        .create_physical_plan()
-        .await?;
-    let schema = plan.schema();
-    let plan = Arc::new(RepartitionExec::try_new(
-        plan,
-        Partitioning::Hash(
-            vec![Arc::new(Column::new_with_schema(HASH_LABEL, &schema)?)],
-            target_partitions,
-        ),
-    )?);
-
-    if config::get_config().common.print_key_sql {
-        log::info!(
-            "{}",
-            config::meta::plan::generate_plan_string(trace_id, plan.as_ref())
-        );
-    }
-
-    let streams = execute_stream_partitioned(plan, ctx)?;
-    let mut tasks = Vec::new();
-    for mut stream in streams {
-        let hash_field_type = hash_field_type.clone();
-        let mut series: HashMap<u64, Vec<Sample>> = HashMap::new();
-        let task: TokioResult = tokio::task::spawn(async move {
-            loop {
-                match stream.try_next().await {
-                    Ok(Some(batch)) => {
-                        let time_values = batch
-                            .column_by_name(TIMESTAMP_COL_NAME)
-                            .unwrap()
-                            .as_any()
-                            .downcast_ref::<Int64Array>()
-                            .unwrap();
-                        let value_values = batch
-                            .column_by_name(VALUE_LABEL)
-                            .unwrap()
-                            .as_any()
-                            .downcast_ref::<Float64Array>()
-                            .unwrap();
-
-                        if hash_field_type == DataType::UInt64 {
-                            let hash_values = batch
-                                .column_by_name(HASH_LABEL)
-                                .unwrap()
-                                .as_any()
-                                .downcast_ref::<UInt64Array>()
-                                .unwrap();
-                            for i in 0..batch.num_rows() {
-                                let timestamp = time_values.value(i);
-                                let hash: u64 = hash_values.value(i);
-                                let entry = series.entry(hash).or_default();
-                                entry.push(Sample::new(timestamp, value_values.value(i)));
-                            }
-                        } else {
-                            let hash_values = batch
-                                .column_by_name(HASH_LABEL)
-                                .unwrap()
-                                .as_any()
-                                .downcast_ref::<StringArray>()
-                                .unwrap();
-                            for i in 0..batch.num_rows() {
-                                let timestamp = time_values.value(i);
-                                let hash: u64 = gxhash::new().sum64(hash_values.value(i));
-                                let entry = series.entry(hash).or_default();
-                                entry.push(Sample::new(timestamp, value_values.value(i)));
-                            }
-                        }
-                    }
-                    Ok(None) => break,
-                    Err(e) => {
-                        log::error!("load samples from datafusion execute stream Error: {e}");
-                        return Err(e);
-                    }
-                }
-            }
-            Ok(series)
-        });
-        tasks.push(task);
-    }
-
-    // collect results
-    let mut metrics: HashMap<u64, RangeValue> = HashMap::new();
-    for task in tasks {
-        let m = task
-            .await
-            .map_err(|e| DataFusionError::Execution(e.to_string()))??;
-        for (hash, samples) in m {
-            metrics.insert(
-                hash,
-                RangeValue {
-                    labels: vec![],
-                    samples,
-                    exemplars: None,
-                    time_window: None,
-                },
-            );
-        }
-    }
-
-    Ok(metrics)
-}
-
-async fn load_exemplars_from_datafusion(
-    trace_id: &str,
-    hash_field_type: &DataType,
-    df: DataFrame,
-) -> Result<HashMap<u64, RangeValue>> {
-    let ctx = Arc::new(df.task_ctx());
-    let target_partitions = ctx.session_config().target_partitions();
-    let plan = df
-        .filter(col(EXEMPLARS_LABEL).is_not_null())?
-        .select_columns(&[HASH_LABEL, EXEMPLARS_LABEL])?
-        .create_physical_plan()
-        .await?;
-    let schema = plan.schema();
-    let plan = Arc::new(RepartitionExec::try_new(
-        plan,
-        Partitioning::Hash(
-            vec![Arc::new(Column::new_with_schema(HASH_LABEL, &schema)?)],
-            target_partitions,
-        ),
-    )?);
-
-    if config::get_config().common.print_key_sql {
-        log::info!(
-            "{}",
-            config::meta::plan::generate_plan_string(trace_id, plan.as_ref())
-        );
-    }
-
-    let streams = execute_stream_partitioned(plan, ctx)?;
-    let mut tasks = Vec::new();
-    for mut stream in streams {
-        let hash_field_type = hash_field_type.clone();
-        let mut series: HashMap<u64, Vec<Arc<Exemplar>>> = HashMap::new();
-        let task: TokioExemplarsResult = tokio::task::spawn(async move {
-            loop {
-                match stream.try_next().await {
-                    Ok(Some(batch)) => {
-                        let exemplars_values = batch
-                            .column_by_name(EXEMPLARS_LABEL)
-                            .unwrap()
-                            .as_any()
-                            .downcast_ref::<StringArray>()
-                            .unwrap();
-                        if hash_field_type == DataType::UInt64 {
-                            let hash_values = batch
-                                .column_by_name(HASH_LABEL)
-                                .unwrap()
-                                .as_any()
-                                .downcast_ref::<UInt64Array>()
-                                .unwrap();
-                            for i in 0..batch.num_rows() {
-                                let hash: u64 = hash_values.value(i);
-                                let exemplar = exemplars_values.value(i);
-                                if let Ok(exemplars) = json::from_str::<Vec<json::Value>>(exemplar)
-                                {
-                                    let entry = series.entry(hash).or_default();
-                                    for exemplar in exemplars {
-                                        if let Some(exemplar) = exemplar.as_object() {
-                                            entry.push(Arc::new(Exemplar::from(exemplar)));
-                                        }
-                                    }
-                                }
-                            }
-                        } else {
-                            let hash_values = batch
-                                .column_by_name(HASH_LABEL)
-                                .unwrap()
-                                .as_any()
-                                .downcast_ref::<StringArray>()
-                                .unwrap();
-                            for i in 0..batch.num_rows() {
-                                let hash: u64 = gxhash::new().sum64(hash_values.value(i));
-                                let exemplar = exemplars_values.value(i);
-                                if let Ok(exemplars) = json::from_str::<Vec<json::Value>>(exemplar)
-                                {
-                                    let entry = series.entry(hash).or_default();
-                                    for exemplar in exemplars {
-                                        if let Some(exemplar) = exemplar.as_object() {
-                                            entry.push(Arc::new(Exemplar::from(exemplar)));
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    Ok(None) => break,
-                    Err(e) => {
-                        log::error!("load exemplars from datafusion execute stream Error: {e}");
-                        return Err(e);
-                    }
-                }
-            }
-            Ok(series)
-        });
-        tasks.push(task);
-    }
-
-    // collect results
-    let mut metrics: HashMap<u64, RangeValue> = HashMap::new();
-    for task in tasks {
-        let m = task
-            .await
-            .map_err(|e| DataFusionError::Execution(e.to_string()))??;
-        for (hash, exemplars) in m {
-            metrics.insert(
-                hash,
-                RangeValue {
-                    labels: vec![],
-                    samples: vec![],
-                    exemplars: Some(exemplars),
-                    time_window: None,
-                },
-            );
-        }
-    }
-
-    Ok(metrics)
-}
-
-/// Prepend the synthetic `__hash__` label when the query asked for raw data
-/// (`query_data`); otherwise return the labels unchanged.
-fn maybe_prepend_hash_label(labels: Labels, hash: u64, query_data: bool) -> Labels {
-    if !query_data {
-        return labels;
-    }
-    let mut with_hash = Vec::with_capacity(labels.len() + 1);
-    with_hash.push(Arc::new(Label {
-        name: HASH_LABEL.to_string(),
-        value: hash.to_string(),
-    }));
-    with_hash.extend(labels);
-    with_hash
-}
-
-/// Aggregations that with no modifier group every series into a single
-/// labelless output series, so the input labels are provably unused.
-const LABEL_DROPPING_AGGS: [u8; 8] = [
-    token::T_SUM,
-    token::T_AVG,
-    token::T_COUNT,
-    token::T_MIN,
-    token::T_MAX,
-    token::T_GROUP,
-    token::T_STDDEV,
-    token::T_STDVAR,
-];
-
-/// Functions that neither read nor create label values — they only transform
-/// per-series samples. Anything label-sensitive (`label_replace`,
-/// `histogram_quantile`, `absent`, ...) must NOT be listed here.
-const LABEL_AGNOSTIC_FUNCS: [&str; 32] = [
-    "rate",
-    "irate",
-    "increase",
-    "delta",
-    "idelta",
-    "deriv",
-    "changes",
-    "resets",
-    "avg_over_time",
-    "min_over_time",
-    "max_over_time",
-    "sum_over_time",
-    "count_over_time",
-    "last_over_time",
-    "stddev_over_time",
-    "stdvar_over_time",
-    "quantile_over_time",
-    "predict_linear",
-    "holt_winters",
-    "abs",
-    "ceil",
-    "floor",
-    "exp",
-    "sqrt",
-    "ln",
-    "log2",
-    "log10",
-    "round",
-    "sgn",
-    "clamp",
-    "clamp_max",
-    "clamp_min",
-];
-
-/// Returns true when the query's root aggregation discards all labels and the
-/// subtree below it never reads label values, so series labels don't need to
-/// be loaded at all.
-fn labels_dropped_at_root(expr: &PromExpr) -> bool {
-    match expr {
-        PromExpr::Aggregate(AggregateExpr {
-            op,
-            expr,
-            param,
-            modifier,
-        }) => {
-            modifier.is_none()
-                && param.is_none()
-                && LABEL_DROPPING_AGGS.contains(&op.id())
-                && subtree_labels_unused(expr)
-        }
-        PromExpr::Paren(ParenExpr { expr }) => labels_dropped_at_root(expr),
-        _ => false,
-    }
-}
-
-fn subtree_labels_unused(expr: &PromExpr) -> bool {
-    match expr {
-        PromExpr::VectorSelector(_) | PromExpr::MatrixSelector(_) => true,
-        PromExpr::NumberLiteral(_) => true,
-        PromExpr::Paren(ParenExpr { expr }) => subtree_labels_unused(expr),
-        PromExpr::Unary(UnaryExpr { expr }) => subtree_labels_unused(expr),
-        PromExpr::Subquery(subquery) => subtree_labels_unused(&subquery.expr),
-        PromExpr::Call(Call { func, args }) => {
-            LABEL_AGNOSTIC_FUNCS.contains(&func.name)
-                && args.args.iter().all(|arg| subtree_labels_unused(arg))
-        }
-        _ => false,
-    }
-}
-
 fn get_offset_modifier(offset: Option<Offset>) -> i64 {
     if let Some(offset) = offset {
         match offset {
@@ -1977,39 +1302,6 @@ mod tests {
 
         fn children(&self) -> &[promql_parser::parser::Expr] {
             &[]
-        }
-    }
-
-    #[test]
-    fn test_labels_dropped_at_root() {
-        let cases = [
-            ("sum(rate(metric[5m]))", true),
-            ("avg(irate(metric[1m]))", true),
-            ("max(abs(metric))", true),
-            ("count(metric)", true),
-            ("(sum(rate(metric[5m])))", true),
-            ("sum(clamp(rate(metric[5m]), 0, 100))", true),
-            // grouping keeps labels
-            ("sum by (region) (rate(metric[5m]))", false),
-            ("sum without (le) (rate(metric[5m]))", false),
-            // root is not an aggregation
-            ("rate(metric[5m])", false),
-            ("sum(rate(metric[5m])) / 2", false),
-            // label-sensitive constructs
-            (
-                "histogram_quantile(0.9, sum by (le) (rate(metric[5m])))",
-                false,
-            ),
-            ("topk(3, rate(metric[5m]))", false),
-            (
-                "sum(label_replace(metric, \"a\", \"$1\", \"b\", \"(.*)\"))",
-                false,
-            ),
-            ("sum(metric_a + metric_b)", false),
-        ];
-        for (query, expected) in cases {
-            let expr = promql_parser::parser::parse(query).unwrap();
-            assert_eq!(labels_dropped_at_root(&expr), expected, "query: {query}");
         }
     }
